requirements_test.txt
coverage==4.4.1
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
codecov>=2.0.0


# Additional test requirements go here
bsdiff4==1.1.4
six==1.11.0
future==0.16.0
paramiko==2.4.0
PyYAML
-e git+https://github.com/jserver/mock-s3.git#egg=0.1
psycopg2
redis
hiredis
django-redis-cache
boto
botocore
django-storages
